Generalized a lot of launching logic into the saas launcher away from the hosted installation script: mm2.py.
